chore(examples): restructure CUDA examples and add a GEMM example
- Refactored the CUDA examples directory for improved clarity and
  modularity so each example is more self-contained.
- Added a new GEMM (General Matrix Multiply) example, including naive
  and tiled kernel implementations, build scripts, and benchmarks.
- The tiled-gemm kernel demonstrates the shared memory usage.
