+++ Cargo.toml
@@ -8,9 +8,14 @@ members = [
 
   "xtask",
 
+  "examples/cuda/vecadd",
+  "examples/cuda/vecadd/kernels",
+  "examples/cuda/gemm",
+  "examples/cuda/gemm/kernels",
+  "examples/cuda/path_tracer",
+  "examples/cuda/path_tracer/kernels",
+
   "examples/optix/*",
-  "examples/cuda/cpu/*",
-  "examples/cuda/gpu/*",
 ]
 
 exclude = [
