--- examples/cuda/cpu/add/build.rs
@@ -1,8 +1,0 @@
-use cuda_builder::CudaBuilder;
-
-fn main() {
-    CudaBuilder::new("../../gpu/add_gpu")
-        .copy_to("../../resources/add.ptx")
-        .build()
-        .unwrap();
-}
