--- examples/cuda/cpu/path_tracer/build.rs
@@ -1,14 +1,0 @@
-use cuda_builder::CudaBuilder;
-
-fn main() {
-    CudaBuilder::new("../../gpu/path_tracer_gpu")
-        .copy_to("../../resources/path_tracer.ptx")
-        .build()
-        .unwrap();
-
-    CudaBuilder::new("../../gpu/path_tracer_gpu")
-        .copy_to("../../resources/path_tracer_optix.ptx")
-        .build_args(&["--features", "optix"])
-        .build()
-        .unwrap();
-}
