+++ examples/cuda/gemm/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "gemm"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+blastoff = { path = "../../../crates/blastoff" }
+cuda_std = { path = "../../../crates/cuda_std" }
+cust = { path = "../../../crates/cust" }
+cust_raw = { path = "../../../crates/cust_raw", features = ["driver"] }
+ndarray = { version = "0.16", features = ["approx"] }
+ndarray-rand = "0.15.0"
+rand = "0.9"
+
+[build-dependencies]
+cuda_builder = { path = "../../../crates/cuda_builder" }
