--- examples/cuda/gpu/add_gpu/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "add_gpu"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-cuda_std = { version = "0.2", path = "../../../../crates/cuda_std" }
-
-[lib]
-crate-type = ["cdylib", "rlib"]
