--- examples/cuda/gpu/add_gpu/src/lib.rs
@@ -1,11 +1,0 @@
-use cuda_std::prelude::*;
-
-#[kernel]
-#[allow(improper_ctypes_definitions, clippy::missing_safety_doc)]
-pub unsafe fn add(a: &[f32], b: &[f32], c: *mut f32) {
-    let idx = thread::index_1d() as usize;
-    if idx < a.len() {
-        let elem = &mut *c.add(idx);
-        *elem = a[idx] + b[idx];
-    }
-}
