--- examples/cuda/gpu/path_tracer_gpu/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "path_tracer_gpu"
-version = "0.1.0"
-edition = "2018"
-
-[dependencies]
-cuda_std = { version = "0.2", path = "../../../../crates/cuda_std" }
-enum_dispatch = "0.3.13"
-gpu_rand = { version = "0.1", path = "../../../../crates/gpu_rand" }
-cust_core = { path = "../../../../crates/cust_core", features=["vek"] }
-optix_device = { path = "../../../../crates/optix_device" }
-
-[lib]
-crate-type = ["cdylib", "rlib"]
-
-[features]
-optix = []
