+++ examples/cuda/path_tracer/kernels/src/hittable.rs
