+++ examples/cuda/path_tracer/kernels/src/lib.rs
