+++ examples/cuda/path_tracer/kernels/src/material.rs
