+++ examples/cuda/path_tracer/kernels/src/math.rs
