+++ examples/cuda/path_tracer/kernels/src/optix.rs
