+++ examples/cuda/path_tracer/kernels/src/render.rs
