+++ examples/cuda/path_tracer/kernels/src/render_kernels.rs
