+++ examples/cuda/path_tracer/kernels/src/scene.rs
