+++ examples/cuda/path_tracer/kernels/src/sphere.rs
