+++ examples/cuda/path_tracer/shaders/image.frag
