+++ examples/cuda/path_tracer/shaders/image.vert
