+++ examples/cuda/vecadd/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "vecadd"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+cust = { path = "../../../crates/cust" }
+nanorand = "0.7"
+
+[build-dependencies]
+cuda_builder = { path = "../../../crates/cuda_builder" }
