+++ examples/cuda/vecadd/kernels/src/lib.rs
@@ -1,0 +1,11 @@
+use cuda_std::prelude::*;
+
+#[kernel]
+#[allow(improper_ctypes_definitions, clippy::missing_safety_doc)]
+pub unsafe fn vecadd(a: &[f32], b: &[f32], c: *mut f32) {
+    let idx = thread::index_1d() as usize;
+    if idx < a.len() {
+        let elem = unsafe { &mut *c.add(idx) };
+        *elem = a[idx] + b[idx];
+    }
+}
